+++ tests/testthat/test-separate-lazy.R
@@ -13,30 +13,37 @@ test_that("basic behavior works", {
     sep = "\\."
   ))
 
-  # TODO: both should take the same input when https://github.com/pola-rs/polars/pull/26060
-  # is released
+  expect_equal_lazy(
+    separate(test_pl, x, into = c("foo", "foo2"), sep = "\\."),
+    separate(test, x, into = c("foo", "foo2"), sep = "\\.")
+  )
+
   expect_equal_lazy(
     separate(test_pl, x, into = c("foo", "foo2"), sep = "."),
-    separate(test, x, into = c("foo", "foo2"), sep = "\\.")
+    suppressWarnings(
+      separate(test, x, into = c("foo", "foo2"), sep = ".")
+    )
   )
 })
 
 test_that("default value for sep works", {
-  # tidypolars-specific: default sep differs from tidyr
-  test <- tibble(x = c(NA, "x y", "x z", "y z"))
+  test <- tibble(x = c(NA, "x y", "x  z", "y   z  u"))
   test_pl <- as_polars_lf(test)
 
-  # TODO: test more extensively regex
-  # https://github.com/pola-rs/polars/issues/4819
-  # https://github.com/pola-rs/polars/pull/26060
   expect_equal_lazy(
-    separate(test_pl, x, into = c("foo", "foo2")) |> pull(foo),
-    c(NA, "x", "x", "y")
+    separate(test_pl, x, into = c("foo", "foo2"), sep = "\\s+"),
+    suppressWarnings(
+      separate(test, x, into = c("foo", "foo2"), sep = "\\s+")
+    )
   )
+})
 
-  expect_equal_lazy(
-    separate(test_pl, x, into = c("foo", "foo2")) |> pull(foo2),
-    c(NA, "y", "z", "z")
+test_that("tidypolars only supports character separator", {
+  test <- pl$LazyFrame(x = c(NA, "x y", "x  z", "y   z"))
+
+  expect_snapshot_lazy(
+    separate(test_pl, x, into = c("foo", "foo2"), sep = 1),
+    error = TRUE
   )
 })